the past 2-3 week's work.
I've spent hours on this work trying to make it look perfect. The backend still needs some work done. particularly the music selection. I've used a HTML 5 audio element and haven't used any javascript code for it. Forgive me for it. But I did use some javascript for the player.php and the meganote.php